refactor(cleanup): moves general types from index to types file, removes some commented-out logs
